Factored the code
Need to add adaptive threshold for color
